move port definitions to a common location
