+++ cmd/kube-apiserver/app/options/BUILD
@@ -19,10 +19,10 @@ go_library(
         "//pkg/api/legacyscheme:go_default_library",
         "//pkg/apis/core:go_default_library",
         "//pkg/cloudprovider/providers:go_default_library",
+        "//pkg/cluster/ports:go_default_library",
         "//pkg/features:go_default_library",
         "//pkg/kubeapiserver/options:go_default_library",
         "//pkg/kubelet/client:go_default_library",
-        "//pkg/master/ports:go_default_library",
         "//pkg/master/reconcilers:go_default_library",
         "//pkg/serviceaccount:go_default_library",
         "//staging/src/k8s.io/apiextensions-apiserver/pkg/apiserver:go_default_library",
