+++ cmd/kube-proxy/app/BUILD
@@ -19,9 +19,9 @@ go_library(
     importpath = "k8s.io/kubernetes/cmd/kube-proxy/app",
     deps = [
         "//pkg/apis/core:go_default_library",
+        "//pkg/cluster/ports:go_default_library",
         "//pkg/features:go_default_library",
         "//pkg/kubelet/qos:go_default_library",
-        "//pkg/master/ports:go_default_library",
         "//pkg/proxy:go_default_library",
         "//pkg/proxy/apis:go_default_library",
         "//pkg/proxy/apis/config:go_default_library",
