+++ pkg/BUILD
@@ -54,6 +54,7 @@ filegroup(
         "//pkg/client/testdata:all-srcs",
         "//pkg/client/tests:all-srcs",
         "//pkg/cloudprovider:all-srcs",
+        "//pkg/cluster/ports:all-srcs",
         "//pkg/controller:all-srcs",
         "//pkg/credentialprovider:all-srcs",
         "//pkg/features:all-srcs",
