+++ pkg/apis/core/validation/BUILD
@@ -22,9 +22,9 @@ go_library(
         "//pkg/apis/core/v1:go_default_library",
         "//pkg/apis/core/v1/helper:go_default_library",
         "//pkg/capabilities:go_default_library",
+        "//pkg/cluster/ports:go_default_library",
         "//pkg/features:go_default_library",
         "//pkg/fieldpath:go_default_library",
-        "//pkg/master/ports:go_default_library",
         "//pkg/security/apparmor:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/api/events/v1beta1:go_default_library",
