+++ pkg/kubelet/apis/config/fuzzer/BUILD
@@ -6,11 +6,11 @@ go_library(
     importpath = "k8s.io/kubernetes/pkg/kubelet/apis/config/fuzzer",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/cluster/ports:go_default_library",
         "//pkg/kubelet/apis/config:go_default_library",
         "//pkg/kubelet/apis/config/v1beta1:go_default_library",
         "//pkg/kubelet/qos:go_default_library",
         "//pkg/kubelet/types:go_default_library",
-        "//pkg/master/ports:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime/serializer:go_default_library",
         "//staging/src/k8s.io/kubelet/config/v1beta1:go_default_library",
