+++ pkg/kubelet/apis/config/fuzzer/fuzzer.go
@@ -28,7 +28,7 @@ import (
 	kubeletconfigv1beta1 "k8s.io/kubernetes/pkg/kubelet/apis/config/v1beta1"
 	"k8s.io/kubernetes/pkg/kubelet/qos"
 	kubetypes "k8s.io/kubernetes/pkg/kubelet/types"
-	"k8s.io/kubernetes/pkg/master/ports"
+	"k8s.io/kubernetes/pkg/cluster/ports"
 )
 
 // Funcs returns the fuzzer functions for the kubeletconfig apis.
