+++ test/e2e/framework/.import-restrictions
@@ -160,7 +160,7 @@ rules:
       - k8s.io/kubernetes/pkg/kubelet/volumemanager/populator
       - k8s.io/kubernetes/pkg/kubelet/volumemanager/reconciler
       - k8s.io/kubernetes/pkg/kubemark
-      - k8s.io/kubernetes/pkg/master/ports
+      - k8s.io/kubernetes/pkg/cluster/ports
       - k8s.io/kubernetes/pkg/probe
       - k8s.io/kubernetes/pkg/probe/exec
       - k8s.io/kubernetes/pkg/probe/http
