+++ test/e2e/node/BUILD
@@ -24,9 +24,9 @@ go_library(
     importpath = "k8s.io/kubernetes/test/e2e/node",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/cluster/ports:go_default_library",
         "//pkg/kubelet/apis/stats/v1alpha1:go_default_library",
         "//pkg/kubelet/runtimeclass/testing:go_default_library",
-        "//pkg/master/ports:go_default_library",
         "//pkg/util/slice:go_default_library",
         "//staging/src/k8s.io/api/batch/v1:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
