+++ test/e2e/storage/vsphere/BUILD
@@ -39,8 +39,8 @@ go_library(
     ],
     importpath = "k8s.io/kubernetes/test/e2e/storage/vsphere",
     deps = [
+        "//pkg/cluster/ports:go_default_library",
         "//pkg/controller/volume/events:go_default_library",
-        "//pkg/master/ports:go_default_library",
         "//staging/src/k8s.io/api/apps/v1:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/api/storage/v1:go_default_library",
